+++ Sources/NnSwiftDataKit/SwiftDataContextError.swift
@@ -1,0 +1,10 @@
+//
+//  SwiftDataContextError.swift
+//  NnSwiftDataKit
+//
+//  Created by Nikolai Nobadi on 11/29/25.
+//
+
+public enum SwiftDataContextError: Error {
+    case noAppGroupAccess
+}